Implement incomplete illumos support.
illumos requires some additional setup compared to other OS, with
ptem(7M) and ldterm(7M) STREAMS modules needing to be pushed prior to
proper operation of the subsidiary device.  While not ideal, this is
currently done in unlockpt() to avoid having to change the API.

While this builds, it doesn't pass the test suite.  illumos again
behaves differently, as where (all?) other OS allow for ioctls on either
the manager or subsidiary file descriptors, illumos requires they be
performed on the subsidiary only.

The current implementation of this module directs all operations to the
manager descriptor, thus all tests fail with EINVAL.